+++ pymedio/dicom.py
@@ -27,35 +27,53 @@ import numpy.typing as npt
 
 try:
     import pydicom
-except (ModuleNotFoundError, ImportError) as imp_exn:
-    msg = f"pydicom must be installed to use {__name__}."
-    raise RuntimeError(msg) from imp_exn
+except ImportError as imp_exn:
+    imp_exn_msg = f"pydicom must be installed to use {__name__}."
+    raise ImportError(imp_exn_msg) from imp_exn
 
-import medio.base as miob
-import medio.exceptions as mioe
-import medio.typing as miot
-import medio.utils as miou
+import pymedio.base as miob
+import pymedio.exceptions as mioe
+import pymedio.typing as miot
+import pymedio.utils as miou
 
 logger = logging.getLogger(__name__)
+
+ORIENTATION_ATOL = 1e-5
+
+T = typing.TypeVar("T")
+
+
+def _all_float_like(seq: typing.Sequence[builtins.float]) -> builtins.bool:
+    return all(isinstance(x, (float, int)) for x in seq)
 
 
 @dataclasses.dataclass(frozen=True)
 class Cosines:
+    # dicom-numpy -> dicom_numpy/combine_slices.py
+    # ITK -> Modules/IO/GDCM/src/itkGDCMImageIO.cxx
     row: npt.NDArray
     column: npt.NDArray
     slice: npt.NDArray
+
+    def __repr__(self) -> builtins.str:
+        return f"Cosines(row={self.row}, column={self.column}, slice={self.slice})"
 
     @classmethod
     def from_orientation(
         cls: typing.Type[Cosines],
         image_orientation: typing.Sequence[builtins.float] | npt.NDArray,
     ) -> Cosines:
-        row_cosine = np.asanyarray(image_orientation[:3], dtype=np.float64)
-        column_cosine = np.asanyarray(image_orientation[3:], dtype=np.float64)
+        if isinstance(image_orientation, np.ndarray):
+            if image_orientation.size != 6 or image_orientation.ndim != 1:
+                raise ValueError("image_orientation must be seq. of 1 dim and len=6.")
+        elif len(image_orientation) != 6 or not _all_float_like(image_orientation):
+            raise ValueError("image_orientation must be seq. floats with len=6.")
+        row_cosine = miou.to_f64(image_orientation[:3])
+        column_cosine = miou.to_f64(image_orientation[3:])
         slice_cosine = np.cross(row_cosine, column_cosine)
         cosines = cls(row_cosine, column_cosine, slice_cosine)
-        cosines.validate()
         cosines.writable(False)
+        cosines.validate()
         return cosines
 
     def validate(self) -> None:
@@ -103,28 +121,36 @@ class Cosines:
 @dataclasses.dataclass(frozen=True)
 class SortedSlices:
     slices: typing.Tuple[pydicom.Dataset, ...]
-    indices: typing.Tuple[builtins.int, ...]
+    indices: typing.Tuple[builtins.int, ...]  # mapping to get original slice order
     positions: typing.Tuple[builtins.float, ...]
     cosines: Cosines
 
+    def __repr__(self) -> builtins.str:
+        return f"SortedSlices(n_slices={len(self)}, cosines={self.cosines!r})"
+
     def __len__(self) -> builtins.int:
-        length = len(self.slices)
-        assert length == len(self.indices)
-        assert length == len(self.positions)
-        return length
+        _len = len(self.slices)
+        exn_msg: typing.List[builtins.str] = []
+        if _len != (ind_len := len(self.indices)):
+            exn_msg.append(f"num slices {_len} != num indices {ind_len}")
+        if _len != (pos_len := len(self.positions)):
+            exn_msg.append(f"num slices {_len} != num positions {pos_len}")
+        if exn_msg:
+            raise RuntimeError(" ".join(exn_msg))
+        return _len
 
     @classmethod
     def from_datasets(
-        cls: typing.Type[SortedSlices], slice_datasets: typing.Sequence[pydicom.Dataset]
+        cls, slice_datasets: typing.Sequence[pydicom.Dataset]
     ) -> SortedSlices:
         """sort list of pydicom datasets into the correct order"""
-        assert slice_datasets, "slice_datasets empty"
-        image_orientation = np.asanyarray(
-            slice_datasets[0].ImageOrientationPatient, dtype=np.float64
-        )
+        if not slice_datasets:
+            raise ValueError("slice_datasets empty")
+        image_orientation = miou.to_f64(slice_datasets[0].ImageOrientationPatient)
         cosines = Cosines.from_orientation(image_orientation)
         cs = cosines.slice
-        positions = [np.dot(cs, d.ImagePositionPatient).item() for d in slice_datasets]
+        imps = (miou.to_f64(sd.ImagePositionPatient) for sd in slice_datasets)
+        positions = (np.dot(cs, imp).item() for imp in imps)
         _sorted = typing.cast(
             typing.Iterable[typing.Tuple[builtins.int, builtins.float]],
             sorted(enumerate(positions), key=operator.itemgetter(1)),
@@ -143,6 +169,7 @@ class SortedSlices:
         *,
         max_nonuniformity: builtins.float = 5e-4,
         fail_outside_max_nonuniformity: builtins.bool = True,
+        missing_slices_cutoff: builtins.float = 1e-1,
     ) -> None:
         if len(self) > 1:
             diffs = np.diff(self.positions)
@@ -152,17 +179,51 @@ class SortedSlices:
                     raise mioe.OutsideMaxNonUniformity(msg)
                 else:
                     warnings.warn(msg)
+            if not np.allclose(diffs, diffs[0], atol=0.0, rtol=missing_slices_cutoff):
+                msg = "There appear to be missing slices."
+                raise mioe.MissingSlicesException(msg)
+
+    def remove_anomalous_slices(
+        self, *, strict_unique: builtins.bool = True
+    ) -> SortedSlices:
+        to_float_tuple = lambda xs: tuple(float(x) for x in xs)  # noqa: E731
+        orientations = [to_float_tuple(s.ImageOrientationPatient) for s in self.slices]
+        if strict_unique:
+            unq_oris, counts = np.unique(orientations, axis=0, return_counts=True)
+            most_common_orientation = unq_oris[np.argmax(counts)]
+        else:
+            approx_unique_orientations = self._approx_unique(orientations)
+            most_common_orientation = approx_unique_orientations[-1]
+        out = [
+            (s, idx, pos)
+            for s, idx, pos, o in self._zip(orientations)
+            if np.allclose(o, most_common_orientation, atol=ORIENTATION_ATOL)
+        ]
+        new_slices, new_indices, new_positions = miou.unzip(out)
+        if (n_removed := (len(self) - len(new_slices))) > 1:
+            warnings.warn(f"{n_removed} anomalous images removed.")
+        elif n_removed < 0:
+            raise RuntimeError("Images added in remove image func. Report error.")
+        new_image_orientation = miou.to_f64(new_slices[0].ImageOrientationPatient)
+        new_cosines = Cosines.from_orientation(new_image_orientation)
+        return SortedSlices(
+            slices=new_slices,
+            indices=new_indices,
+            positions=new_positions,
+            cosines=new_cosines,
+        )
 
     @functools.cached_property
     def patient_position(self) -> npt.NDArray:
-        return np.asanyarray(self.slices[0].ImagePositionPatient, dtype=np.float64)
+        return miou.to_f64(self.slices[0].ImagePositionPatient)
 
     @functools.cached_property
     def slice_spacing(self) -> builtins.float:
         spacing: builtins.float
         if len(self) > 1:
-            slice_positions_diffs = np.diff(sorted(self.positions))
-            spacing = float(np.median(slice_positions_diffs).item())
+            slice_positions_diffs = np.diff(np.sort(self.positions))
+            # avg. b/c that's what ITK seems to use, so use for consistency
+            spacing = float(np.mean(slice_positions_diffs).item())
         elif len(self) == 1:
             spacing = float(getattr(self.slices[0], "SpacingBetweenSlices", 0))
         else:
@@ -178,24 +239,62 @@ class SortedSlices:
         transform[:3, 1] = self.cosines.column * row_spacing
         transform[:3, 2] = self.cosines.slice * slice_spacing
         transform[:3, 3] = self.patient_position
-        _flipxy_44 = miou.flipxy_44()
-        transform_ras: npt.NDArray = np.dot(_flipxy_44, transform)
+        transform_ras: npt.NDArray = np.dot(miou.flipxy_44(), transform)
         return transform_ras
+
+    @staticmethod
+    def _approx_unique(
+        values: typing.Sequence[T],
+        *,
+        atol: builtins.float = ORIENTATION_ATOL,
+    ) -> typing.Tuple[T, ...]:
+        # TODO: improve computational efficiency
+        # TODO: fix bad init -> bad result
+        if not values:
+            return tuple()
+        approx_unique: typing.Dict[T, builtins.int] = dict()
+        for val in values:
+            min_dist = np.inf
+            min_dist_val = None
+            for target_val in approx_unique.keys():
+                np_val = miou.to_f64(typing.cast(npt.ArrayLike, val))
+                np_tgt_val = miou.to_f64(typing.cast(npt.ArrayLike, target_val))
+                if np.allclose(np_val, np_tgt_val, atol=atol):
+                    dist = np.linalg.norm(np_val - np_tgt_val).item()
+                    if dist < min_dist:
+                        min_dist = dist
+                        min_dist_val = target_val
+            if min_dist_val is None:
+                approx_unique[val] = 1
+            else:
+                approx_unique[min_dist_val] += 1
+        approx_unq_arrs, _ = miou.unzip(
+            sorted(
+                ((arr, count) for arr, count in approx_unique.items()),
+                key=operator.itemgetter(1),
+            )
+        )
+        return approx_unq_arrs
+
+    def _zip(
+        self, *args: typing.Iterable[typing.Any]
+    ) -> typing.Iterable[typing.Tuple[typing.Any, ...]]:
+        return zip(self.slices, self.indices, self.positions, *args)
 
 
 @dataclasses.dataclass(frozen=True)
 class DICOMDir:
     slices: typing.Tuple[pydicom.Dataset, ...]
     positions: typing.Tuple[builtins.float, ...]
-    spacing: builtins.float
+    slice_spacing: builtins.float
     affine: npt.NDArray
     paths: typing.Tuple[miot.PathLike, ...] | None = None
 
     def __len__(self) -> builtins.int:
-        length = len(self.slices)
-        if self.paths is not None:
-            assert length == len(list(self.paths))
-        return length
+        _len = len(self.slices)
+        if self.paths is not None and _len != (path_len := len(self.paths)):
+            raise RuntimeError(f"num slices {_len} != num paths {path_len}")
+        return _len
 
     @classmethod
     def from_datasets(
@@ -215,17 +314,19 @@ class DICOMDir:
             max_nonuniformity=max_nonuniformity,
             fail_outside_max_nonuniformity=fail_outside_max_nonuniformity,
         )
+        if remove_anomalous_images:
+            sorted_slices = sorted_slices.remove_anomalous_slices()
         positions = tuple(sorted(sorted_slices.positions))
+        idxs = sorted_slices.indices
         dicom_dir = cls(
             slices=sorted_slices.slices,
             positions=positions,
-            spacing=sorted_slices.slice_spacing,
+            slice_spacing=sorted_slices.slice_spacing,
             affine=sorted_slices.affine,
-            paths=None if paths is None else tuple(paths),
+            paths=None if paths is None else tuple(paths[i] for i in idxs),
         )
-        if remove_anomalous_images:
-            dicom_dir = dicom_dir.remove_anomalous_image_paths()
         dicom_dir.writable(False)
+        dicom_dir.validate()
         return dicom_dir
 
     @classmethod
@@ -237,17 +338,18 @@ class DICOMDir:
         fail_outside_max_nonuniformity: builtins.bool = True,
         remove_anomalous_images: builtins.bool = True,
         defer_size: builtins.str | builtins.int | None = "1 KB",
+        extension: builtins.str = ".dcm",
     ) -> DICOMDir:
         paths: typing.Tuple[miot.PathLike, ...]
         if (
             isinstance(dicom_path, (builtins.str, pathlib.Path))
             and (_dcm_dir := pathlib.Path(dicom_path)).is_dir()
         ):
-            paths = tuple(sorted(_dcm_dir.glob("*.dcm")))
+            paths = tuple(sorted(_dcm_dir.glob(f"*{extension}")))
         elif (
             not isinstance(dicom_path, (builtins.str, pathlib.Path))
             and miou.is_iterable(dicom_path)
-            and all(str(p).endswith(".dcm") for p in dicom_path)  # type: ignore[union-attr]  # noqa: E501
+            and all(str(p).endswith(f".{extension}") for p in dicom_path)  # type: ignore[union-attr]  # noqa: E501
         ):
             paths = tuple(dicom_path)  # type: ignore[arg-type]
         else:
@@ -309,46 +411,25 @@ class DICOMDir:
             raise mioe.DicomImportException(msg)
         return datasets
 
-    def remove_anomalous_image_paths(self) -> DICOMDir:
-        orientations = [tuple(img.ImageOrientationPatient) for img in self.slices]
-        unique_orientations, counts = np.unique(
-            orientations, axis=0, return_counts=True
-        )
-        most_common_orientation = tuple(unique_orientations[np.argmax(counts)])
-        paths = ([None] * len(self)) if (no_paths := self.paths is None) else self.paths
-        out = [
-            (img, pos, path)
-            for img, pos, path, o in zip(
-                self.slices, self.positions, paths, orientations
-            )
-            if o == most_common_orientation
-        ]
-        new_images, new_positions, new_image_paths = miou.unzip(out)
-        return DICOMDir(
-            slices=new_images,
-            positions=new_positions,
-            spacing=self.spacing,
-            affine=self.affine,
-            paths=None if no_paths else new_image_paths,
-        )
-
     def validate(self) -> None:
         invariant_properties = frozenset(
-            [
-                "Modality",
-                "SOPClassUID",
-                "SeriesInstanceUID",
-                "Rows",
-                "Columns",
-                "SamplesPerPixel",
-                "PixelSpacing",
-                "PixelRepresentation",
+            (
                 "BitsAllocated",
-            ]
+                "Columns",
+                "Modality",
+                "PixelRepresentation",
+                "PixelSpacing",
+                "Rows",
+                "SamplesPerPixel",
+                "SeriesInstanceUID",
+                "SOPClassUID",
+            )
         )
         for property_name in invariant_properties:
             self._slice_attribute_equal(property_name)
-        self._slice_ndarray_attribute_almost_equal("ImageOrientationPatient", atol=1e-5)
+        self._slice_attribute_almost_equal(
+            "ImageOrientationPatient", atol=ORIENTATION_ATOL
+        )
 
     def writable(self, value: builtins.bool, /) -> None:
         self.affine.flags.writeable = value
@@ -362,7 +443,7 @@ class DICOMDir:
                 msg += f"'{property_name}': {value} != {initial_value}"
                 raise mioe.DicomImportException(msg)
 
-    def _slice_ndarray_attribute_almost_equal(
+    def _slice_attribute_almost_equal(
         self,
         property_name: builtins.str,
         *,
@@ -389,7 +470,7 @@ class DICOMDir:
         return result
 
 
-class DICOMImage(miob.ImageBase):
+class DICOMImage(miob.BasicImage):
     @classmethod
     def from_dicomdir(
         cls: typing.Type[DICOMImage],
@@ -415,12 +496,14 @@ class DICOMImage(miob.ImageBase):
         rescale: builtins.bool | None = None,
         rescale_dtype: npt.DTypeLike = np.float32,
         order: typing.Literal["F", "C"] | None = None,
+        extension: builtins.str = ".dcm",
     ) -> DICOMImage:
         dicomdir = DICOMDir.from_path(
             dicom_path,
             max_nonuniformity=max_nonuniformity,
             fail_outside_max_nonuniformity=fail_outside_max_nonuniformity,
             remove_anomalous_images=remove_anomalous_images,
+            extension=extension,
         )
         return cls.from_dicomdir(
             dicomdir, rescale=rescale, rescale_dtype=rescale_dtype, order=order
@@ -498,7 +581,9 @@ class DICOMImage(miob.ImageBase):
                 msg = f"If order given, must be either 'F' or 'C'. Got {order}."
                 raise ValueError(msg)
 
-        assert voxels.dtype == voxels_dtype
+        if voxels.dtype != voxels_dtype:
+            exn_msg = f"voxels.dtype {voxels.dtype} != requested dtype {voxels_dtype}"
+            raise RuntimeError(exn_msg)
         return voxels
 
     @staticmethod